LPS-79753 When add one document, in the same thread or in the same method, for the same AutoDeleteFileInputStream object, AutoDeleteFileInputStream.close() may execute more than once. This will let _file.deleteOnExit() execute and it will cause memory leak.

+++ portal-kernel/src/com/liferay/portal/kernel/io/AutoDeleteFileInputStream.java
@@ -19,6 +19,8 @@ import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 
+import java.nio.channels.FileChannel;
+
 /**
  * @author Shuyang Zhou
  */
@@ -32,6 +34,12 @@ public class AutoDeleteFileInputStream extends FileInputStream {
 
 	@Override
 	public void close() throws IOException {
+		FileChannel channel = super.getChannel();
+
+		if (!channel.isOpen()) {
+			return;
+		}
+
 		super.close();
 
 		if (!_file.delete()) {
